feat: refactor cli and printer
